* TIM_ex: write_TIM_CCR CCR_val needs to be 32b (for 32b timers TIM2 & TIM5) * PWM: set_PWM_Duty_Scaled duty & scale parameters needs to be 32b (for 32b timers TIM2 & TIM5) + computation cast to uint64_t to operate correctly * PWM_IC: Removed active channel 5 & 6 in init_PWM_IC (seems those channels, when available cannot be used for Input Capture)

+++ PWM.c
@@ -15,15 +15,15 @@
 /*******************/
 /*** PWM DRIVING ***/
 /*******************/
-HAL_StatusTypeDef NONNULL__ set_PWM_Duty_Scaled(const TIM_HandleTypeDef * const pTim, const uint32_t chan, const uint16_t duty, const uint16_t scale)
+HAL_StatusTypeDef NONNULL__ set_PWM_Duty_Scaled(const TIM_HandleTypeDef * const pTim, const uint32_t chan, const uint32_t duty, const uint32_t scale)
 {
-	uint16_t tmp;
+	uint32_t tmp;
 
 	if (!scale)			{ return HAL_ERROR; }	// Division by 0
 
 	if (duty >= scale)	{ tmp = pTim->Instance->ARR + 1; }	// +1 To achieve real 100% duty cycle
 	else if (duty == 0)	{ tmp = 0; }
-	else				{ tmp = (duty * pTim->Instance->ARR) / scale; }
+	else				{ tmp = ((uint64_t) duty * pTim->Instance->ARR) / scale; }
 
 	return write_TIM_CCR(pTim, chan, tmp);
 }
@@ -68,6 +68,7 @@ HAL_StatusTypeDef NONNULL__ init_PWM_Chan(TIM_HandleTypeDef * const pTim, const 
 
 
 /****************************************************************/
+
 
 /*****************/
 /*** LOGIC PWM ***/
